FinishTeam

+++ MyGenshin/Server/GameServer/GameServer/Utils/Time.cs
@@ -140,10 +140,10 @@ class Time
         lastTick = _ticks;
     }
 
-    public static long GetTimeStamp()
+    public static int GetTimeStamp()
     {
         TimeSpan ts = DateTime.UtcNow - new DateTime(1970, 1, 1, 0, 0, 0, 0);
-        return Convert.ToInt64(ts.TotalSeconds);
+        return Convert.ToInt32(ts.TotalSeconds);
     }
 
     public static long GetMillTimeStamp()
